`0 *` is better-defined	 than `0 &`
Multiplication is definitely defined in terms of value.
Bitwise-and is defined in terms of bit representation,
and if you know enough to know that you should really
worry about that difference, then it takes more
thinking and reading to check and prove if 0 &` will
always make things zero, without producing trap values
or whatever.

Anyway, grade schoolers know that multiplying by zero
is zero. It's a simpler, more accessible operation.